pyproject.toml
[project]
name = "ceur-graph"
version = "0.1.0"
description = "API for ceur graph entities"
readme = "README.md"
license= "Apache-2.0"
authors = [
    { name = "tholzheim", email = "tim.holzheim@rwth-aachen.de" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "sparqlwrapper>=2.0.0",
    "wikibaseintegrator>=0.12.12",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.7",
]
